refactor: extract app factory
